requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
yt-dlp==2023.11.16
pytest==7.4.3
pytest-asyncio==0.21.1
respx==0.20.1
httpx==0.25.2
python-multipart
openpyxl>=3.1.0
reportlab>=4.0.0
openai-whisper>=20231117
openai>=1.0.0
transformers>=4.35.0
torch>=2.0.0